fixed csv formatting, can now enter multiple data points

+++ test.py
@@ -6,66 +6,60 @@ import os
 import sys
 import argparse
 import pandas as pd
-
 path = 'list.csv'
+
+if os.path.exists(path):
+    csv_exists = True
+else:
+    csv_exists = False
 
 parser = argparse.ArgumentParser(description='build and alter database')
 parser.add_argument('-d', '--delete', action='store_true')
 args = parser.parse_args()
 
 if args.delete == True:
-    print('Resetting CSV file')
-    os.remove(path)
+    if csv_exists:
+        print('Resetting CSV file')
+        os.remove(path)
+        csv_exists = False
+    else:
+        print('Can\'t delete file that does not exist, exiting')
+        exit()
+
+
 
 dict_toAdd = {'Name': [], 'Year Formed':[]}
 
 #check if file exists
 #check if value already exists
-if os.path.exists('list.csv'):
-    print('file exists')
+if csv_exists:
     df_original = pd.read_csv(path)
-    print(df_original)
 else:
     print('creating new csv')
     df_original = pd.DataFrame(dict_toAdd)
-    print(df_original)
 
-name_toAdd = input('Please input a name to add: ')
-if name_toAdd not in df_original.Name.values:
-    print('Name: %s is not in file' % name_toAdd)
-    dict_toAdd['Name'].append(name_toAdd)
-    year_formed = int(input('Please enter and integer for the year formed: '))
-    dict_toAdd['Year Formed'].append(year_formed)
-    print(dict_toAdd)
-else:
-    print('Name: %s already in file. Exiting...' % name_toAdd)
-    exit()
+continue_input = True
+
+while continue_input:
+    name_toAdd = input('Please input a name to add: ')
+    if name_toAdd not in df_original.Name.values:
+        print('Name: %s is not in file' % name_toAdd)
+        dict_toAdd['Name'].append(name_toAdd)
+        year_formed = int(input('Please enter and integer for the year formed: '))
+        dict_toAdd['Year Formed'].append(year_formed)
+        #print(dict_toAdd)
+    else:
+        print('Name: %s already in file.\n' % name_toAdd)
+    flag = input('Enter more information? --type N to exit--')
+    if flag == 'N':
+        continue_input = False
+        
 
 df_toAdd = pd.DataFrame(dict_toAdd)
-print(df_toAdd)
+print('\nDataFrame to add to original file \n ---------------------------------- \n %s \n' % df_toAdd)
 ret = pd.concat([df_original, df_toAdd])
-print(ret)
-#ret.reset_index(drop=True)
-print(ret)
+ret['Year Formed'] = ret['Year Formed'].astype(int)
 
-csv_data = ret.to_csv(path)
-#with open('list.csv', 'a') as fp:
-#    fp.writelines(value + '\n' for value in missing)
 
-#df = pd.DataFrame(my_df)
-#my_df = pd.Series(header)
-#csv_data = my_df.to_csv('list.csv')
-#with open('list.csv', 'w') as fp:
-#    fp.writelines(header)    
-# #n_args = len(sys.argv)
-    #if n_args == 1:
-    #    print('please include name')
-    #    sys.exit()        
-    #name = sys.argv[1]
-    #print(name)
-    #if n_args == 3:
-    #    year = sys.argv[2]
-    
-    
-    #with open('list.csv', 'r') as fp:
-    #    s = fp.read()
+#print('\nFinal DataFrame \n ---------------------------------- \n %s \n\n' % ret)
+csv_data = ret.to_csv(path, index=False)
